// spark-event-ui/src/app/rest/model/repeatedLongestSqlExecsPerCallSiteDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CallSiteTemplatedStringDTO } from './callSiteTemplatedStringDTO';
import { SparkSQLExecutionEventSummaryDTO } from './sparkSQLExecutionEventSummaryDTO';


export interface RepeatedLongestSqlExecsPerCallSiteDTO { 
    callSite?: CallSiteTemplatedStringDTO;
    count?: number;
    durationSum?: number;
    topSqlExecs?: Array<SparkSQLExecutionEventSummaryDTO>;
}