// spark-event-ui/src/app/rest/model/repeatedLongestTopLevelJobExecsPerCallSiteDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SparkTopLevelJobExecEventSummaryDTO } from './sparkTopLevelJobExecEventSummaryDTO';
import { CallSiteTemplatedStringDTO } from './callSiteTemplatedStringDTO';


export interface RepeatedLongestTopLevelJobExecsPerCallSiteDTO { 
    callSite?: CallSiteTemplatedStringDTO;
    count?: number;
    durationSum?: number;
    topJobsExecs?: Array<SparkTopLevelJobExecEventSummaryDTO>;
}